Fix Solcast loading for double azimuth's. Don't load device list for logged out users.

+++ Energy Stats Core/Models/SolcastSettings.swift
@@ -22,14 +22,14 @@ public struct SolcastSite: Codable, Equatable {
     public let resourceId: String
     public let lng: Double
     public let lat: Double
-    public let azimuth: Int
+    public let azimuth: Double
     public let tilt: Double
     public let lossFactor: Double?
     public let acCapacity: Double
     public let dcCapacity: Double?
     public let installDate: Date?
 
-    public init(name: String, resourceId: String, lng: Double, lat: Double, azimuth: Int, tilt: Double, lossFactor: Double?, acCapacity: Double, dcCapacity: Double?, installDate: Date?) {
+    public init(name: String, resourceId: String, lng: Double, lat: Double, azimuth: Double, tilt: Double, lossFactor: Double?, acCapacity: Double, dcCapacity: Double?, installDate: Date?) {
         self.name = name
         self.resourceId = resourceId
         self.lng = lng
